+++ tests/test_parse.py
@@ -24,21 +24,21 @@ def test_fails():
 
 
 def test_json():
-    assert Model.parse_raw('{"a": 12, "b": 8}') == Model.construct(a=12, b=8)
+    assert Model.parse_raw('{"a": 12, "b": 8}') == Model(a=12, b=8)
 
 
 def test_json_ct():
-    assert Model.parse_raw('{"a": 12, "b": 8}', content_type='application/json') == Model.construct(a=12, b=8)
+    assert Model.parse_raw('{"a": 12, "b": 8}', content_type='application/json') == Model(a=12, b=8)
 
 
 def test_pickle_ct():
     data = pickle.dumps(dict(a=12, b=8))
-    assert Model.parse_raw(data, content_type='application/pickle', allow_pickle=True) == Model.construct(a=12, b=8)
+    assert Model.parse_raw(data, content_type='application/pickle', allow_pickle=True) == Model(a=12, b=8)
 
 
 def test_pickle_proto():
     data = pickle.dumps(dict(a=12, b=8))
-    assert Model.parse_raw(data, proto=Protocol.pickle, allow_pickle=True) == Model.construct(a=12, b=8)
+    assert Model.parse_raw(data, proto=Protocol.pickle, allow_pickle=True) == Model(a=12, b=8)
 
 
 def test_pickle_not_allowed():
@@ -64,22 +64,22 @@ def test_bad_proto():
 def test_file_json(tmpdir):
     p = tmpdir.join('test.json')
     p.write('{"a": 12, "b": 8}')
-    assert Model.parse_file(str(p)) == Model.construct(a=12, b=8)
+    assert Model.parse_file(str(p)) == Model(a=12, b=8)
 
 
 def test_file_json_no_ext(tmpdir):
     p = tmpdir.join('test')
     p.write('{"a": 12, "b": 8}')
-    assert Model.parse_file(str(p)) == Model.construct(a=12, b=8)
+    assert Model.parse_file(str(p)) == Model(a=12, b=8)
 
 
 def test_file_pickle(tmpdir):
     p = tmpdir.join('test.pkl')
     p.write_binary(pickle.dumps(dict(a=12, b=8)))
-    assert Model.parse_file(str(p), allow_pickle=True) == Model.construct(a=12, b=8)
+    assert Model.parse_file(str(p), allow_pickle=True) == Model(a=12, b=8)
 
 
 def test_file_pickle_no_ext(tmpdir):
     p = tmpdir.join('test')
     p.write_binary(pickle.dumps(dict(a=12, b=8)))
-    assert Model.parse_file(str(p), content_type='application/pickle', allow_pickle=True) == Model.construct(a=12, b=8)
+    assert Model.parse_file(str(p), content_type='application/pickle', allow_pickle=True) == Model(a=12, b=8)