+++ HISTORY.rst
@@ -3,8 +3,13 @@
 History
 -------
 
-v0.19.1 (unreleased)
+v0.20.0 (unreleased)
 ....................
+* **breaking change** (maybe): more sophisticated argument parsing for validators, any subset of
+  ``values``, ``config`` and ``field`` is now permitted, eg. ``(cls, value, field)``,
+  however the variadic key word argument ("``**kwargs``") **must** be called ``kwargs``, #388 by @samuelcolvin
+* Adds ``skip_defaults`` argument to ``BaseModel.dict()`` to allow skipping of fields that were not
+  explicitly set, #389 by @dgasmith
 * Fix ``extra`` behaviour for multiple inheritance/mix-ins, #394 by @YaraslauZhylko
 
 v0.19.0 (2019-02-04)